Extract workspace run notification boundary
