+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -9,7 +9,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `e2e:playwright` | 48 | 7465 | 100 | A+ | `revert.spec.ts` A+<br>`interaction-audit-targets.ts` A+<br>`interaction-audit-fixtures.spec.ts` A+ |
 | `scripts` | 23 | 4114 | 99 | A+ | `native-desktop-smoke.sh` A+<br>`smoke.sh` A+<br>`packaged_window.py` A+ |
 | `source:QuillCodeAgent` | 54 | 6879 | 100 | A+ | `TrustedRouterOAuth.swift` A+<br>`FlailDetector.swift` A+<br>`AgentTypes.swift` A+ |
-| `source:QuillCodeApp` | 432 | 48935 | 100 | A+ | `WorkspaceModelComposer.swift` A+<br>`WorkspaceSurface.swift` A+<br>`QuillCodeThreadSidebarSurface.swift` A+ |
+| `source:QuillCodeApp` | 433 | 48962 | 100 | A+ | `WorkspaceSurface.swift` A+<br>`QuillCodeThreadSidebarSurface.swift` A+<br>`WorkspaceMemoriesSurface.swift` A+ |
 | `source:QuillCodeCore` | 31 | 2694 | 99 | A+ | `ModelInfo.swift` A+<br>`AppConfig.swift` A+<br>`AutomationModels.swift` A+ |
 | `source:QuillCodePersistence` | 6 | 354 | 100 | A+ | `ConfigStore.swift` A+<br>`SecretStore.swift` A+<br>`ThreadStore.swift` A+ |
 | `source:QuillCodeSafety` | 7 | 1113 | 99 | A+ | `Safety.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+<br>`StaticSafetyDownloadPolicy.swift` A+ |
@@ -33,7 +33,6 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | Grade | Score | Lines | Module | File | Main Issues |
 | --- | ---: | ---: | --- | --- | --- |
 | A+ | 96 | 576 | `test:QuillCodeCoreTests` | `Tests/QuillCodeCoreTests/CoreModelTests.swift` | 29 lines >120 chars |
-| A+ | 96 | 431 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelComposer.swift` | watch file size (431 lines); 2 lines >120 chars |
 | A+ | 96 | 371 | `scripts` | `scripts/native-desktop-smoke.sh` | 22 lines >120 chars; 1 lines >160 chars |
 | A+ | 96 | 339 | `source:QuillCodeAgent` | `Sources/QuillCodeAgent/TrustedRouterOAuth.swift` | 2 lines >120 chars; 29 public declarations |
 | A+ | 96 | 248 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceSurface.swift` | 1 lines >120 chars; 21 public declarations |
@@ -72,6 +71,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 97 | 206 | `scripts` | `scripts/native_click_probe_contracts/packaged_window.py` | 13 lines >120 chars; 1 lines >160 chars |
 | A+ | 97 | 205 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/Safety.swift` | 19 public declarations |
 | A+ | 97 | 204 | `source:QuillCodeCore` | `Sources/QuillCodeCore/AutomationModels.swift` | 36 public declarations |
+| A+ | 97 | 190 | `source:QuillCodeTools` | `Sources/QuillCodeTools/WorkspaceFileIndexer.swift` | 15 public declarations |
 
 ## Every File
 
@@ -505,7 +505,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 49 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelBrowserSession.swift` | no automated issues |
 | A+ | 100 | 112 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelBrowserSnapshots.swift` | no automated issues |
 | A+ | 99 | 169 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelCatalogSurfaceBuilder.swift` | 1 lines >120 chars |
-| A+ | 96 | 431 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelComposer.swift` | watch file size (431 lines); 2 lines >120 chars |
+| A+ | 99 | 346 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelComposer.swift` | 1 lines >120 chars |
 | A+ | 100 | 84 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelConfiguration.swift` | no automated issues |
 | A+ | 99 | 52 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelContext.swift` | 1 lines >120 chars |
 | A+ | 100 | 197 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelContextContinuations.swift` | no automated issues |
@@ -518,6 +518,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 99 | 67 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelProjectInit.swift` | 1 lines >120 chars |
 | A+ | 100 | 192 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelProjects.swift` | no automated issues |
 | A+ | 100 | 212 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelReview.swift` | no automated issues |
+| A+ | 100 | 112 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelRunNotifications.swift` | no automated issues |
 | A+ | 100 | 157 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelSidebar.swift` | no automated issues |
 | A+ | 100 | 170 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelTerminal.swift` | no automated issues |
 | A+ | 100 | 93 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceModelThreadFollowUpAutomations.swift` | no automated issues |
