+++ docs/DECISIONS.md
@@ -577,3 +577,4 @@
 - Activity source commands preserve source location as typed command data, not renderer-local string conventions. Path-only Open/Edit command IDs remain stable, while diagnostic source actions may use `activity-source-*-line:<line>:<path>` so `file:line` survives Activity rendering, command parsing, model fallback, and desktop routing. Native editor opening is a desktop-only convenience for readable local files inside the active workspace; remote projects, missing files, directories, absolute paths, and workspace escapes must fall back to the normal model/tool-card path.
 - Instruction diagnostic keep-side fixes are intentionally narrow and deterministic. Activity may offer direct patch actions only for two-reference semantic conflicts where the referenced path is a safe generated patch target and the loaded instruction line still matches the diagnostic excerpt. The command removes the opposite conflicting line via `host.apply_patch` and refreshes project context; duplicate-scope and nested-override diagnostics stay manual until a merge strategy can preserve user intent.
 - GitHub pull request listing is a first-class read-only PR tool, not a shell workaround. `host.git.pr.list` owns optional state and limit arguments, `GitHubPullRequestInputValidator` normalizes/bounds those arguments for local execution, SSH Remote command planning, slash parsing, and tests, and command/palette/slash surfaces should expose `/pr list` before label commands for `/pr l` discoverability.
+- Run-finished notifications are a workspace-model boundary separate from composer send orchestration. `WorkspaceModelComposer` owns submit/resume, send lifecycle, progress, and terminal outcome handling; `WorkspaceModelRunNotifications` owns notification gating, verification-action dispatch, and shell verification fallback. Keep future "notify me when done" behavior in the notification extension so the composer does not grow another cross-cutting concern.
